# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /assets:
    get:
      operationId: getAssets
      tags: [assets]
      summary: List all assets from Microsoft Lists
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search by asset number, manufacturer, or model
        - name: status
          in: query
          required: false
          schema:
            type: string
          description: Filter by status
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createAsset
      tags: [assets]
      summary: Add a new asset to Microsoft Lists
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Asset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /assets/low-stock:
    get:
      operationId: getLowStockProducts
      tags: [assets]
      summary: List Rational product combinations with exactly one available AMT asset
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Low-stock product combinations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LowStockProduct"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get a single asset
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Asset detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateAsset
      tags: [assets]
      summary: Update an asset (e.g. change status)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "200":
          description: Updated asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /assets/stats:
    get:
      operationId: getAssetStats
      tags: [assets]
      summary: Get asset counts grouped by status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Asset stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetStats"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
      required:
        - id
        - username
        - displayName
    Asset:
      type: object
      properties:
        id:
          type: string
        assetNumber:
          type: string
        status:
          type: string
        manufacturer:
          type: string
        model:
          type: string
        size:
          type: string
        fuel:
          type: string
        features:
          type: string
        datePurchased:
          type: string
        dateRetested:
          type: string
        needsRetest:
          type: boolean
        supplier:
          type: string
        customer:
          type: string
        createdAt:
          type: string
        modifiedAt:
          type: string
      required:
        - id
        - assetNumber
        - status
        - dateRetested
        - needsRetest
    AssetInput:
      type: object
      properties:
        assetNumber:
          type: string
        status:
          type: string
        manufacturer:
          type: string
        model:
          type: string
        size:
          type: string
        fuel:
          type: string
        features:
          type: string
        datePurchased:
          type: string
        supplier:
          type: string
        customer:
          type: string
      required:
        - assetNumber
        - status
    AssetStats:
      type: object
      properties:
        total:
          type: integer
        byStatus:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
              count:
                type: integer
            required:
              - status
              - count
      required:
        - total
        - byStatus
...
    LowStockProduct:
      type: object
      properties:
        key:
          type: string
        brand:
          type: string
        model:
          type: string
        shopifySize:
          type: string
        amtSize:
          type: string
        fuel:
          type: string
        availableCount:
          type: integer
        assetNumbers:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/LowStockRecommendation"
      required:
        - key
        - brand
        - model
        - shopifySize
        - amtSize
        - fuel
        - availableCount
        - assetNumbers
        - recommendations
    LowStockRecommendation:
      type: object
      properties:
        assetNumber:
          type: string
        status:
          type: string
      required: [assetNumber, status]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error